chore: fixed merge conflicts and refactored the code
