+++ src/main/scala/de/htwg/controller/GameController.scala
@@ -18,8 +18,6 @@ class GameController @Inject() (
   inputHandler.attachSession(session)
 
   private var currentState: GameState = initialState
-  override def getBoard: Board = session.board
-  override def isRedTurn: Boolean = session.isRedTurn
 
   override def startGame(): Unit = {
     notifyObservers(StartGame())
