+++ src/main/scala/de/htwg/controller/IController.scala
@@ -5,8 +5,6 @@ import de.htwg.util.Observable
 
 trait IController extends Observable[GameEvent] {
   def startGame(): Unit
-  def getBoard: Board
-  def isRedTurn: Boolean
 }
