+++ src/main/scala/de/htwg/controller/inputhandler/impl/TuiInputHandler.scala
@@ -3,11 +3,12 @@ package de.htwg.controller.inputhandler.impl
 import de.htwg.controller.inputhandler.InputHandler
 import de.htwg.model.GameSession
 
+import scala.compiletime.uninitialized
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
 class TuiInputHandler extends InputHandler {
-    private var gameSession: GameSession = _
+    private var gameSession: GameSession = uninitialized
 
     override def attachSession(session: GameSession): Unit =
         gameSession = session
