+++ src/main/scala/de/htwg/view/gui/GuiView.scala
@@ -1,509 +1,469 @@
 package de.htwg.view.gui
 
-import de.htwg.controller.inputhandler.InputHandler
 import de.htwg.controller.*
+import de.htwg.controller.inputhandler.InputHandler
 import de.htwg.model.*
-import de.htwg.model.Board.*
 import de.htwg.util.Observer
+import scalafx.application.{JFXApp3, Platform}
+import scalafx.scene.Scene
+import scalafx.scene.canvas.Canvas
+import scalafx.scene.layout.{BorderPane, HBox, StackPane, VBox}
+import scalafx.geometry.{Insets, Pos}
+import scalafx.scene.paint.{Color, CycleMethod, LinearGradient, Stop}
+import scalafx.scene.control.{Alert, Button, Label}
+import scalafx.scene.control.Alert.AlertType
+import scalafx.Includes.*
+import scalafx.animation.*
+import scalafx.util.Duration
 
-import java.awt.{BasicStroke, Color, Font, Graphics2D, RenderingHints}
-import javax.swing.JOptionPane
-import scala.swing.*
-import scala.swing.MenuBar.NoMenuBar.{listenTo, reactions}
-import scala.swing.event.*
+import scala.compiletime.uninitialized
+import de.htwg.model.Board.{Board, BoardBuilder}
 
 class GuiView(inputHandler: InputHandler) extends Observer[GameEvent] {
 
-  // State management
-  private var currentBoard: Option[Board] = None
-  private var isRedTurn: Boolean = true
-  private var selectedPiece: Option[(Int, Int)] = None // (Internal Row, Internal Col)
-  private var validMoves: List[(Int, Int, Boolean)] = Nil // (Internal Row, Internal Col, isJump)
-  private var statusMessage: String = "Welcome to Checkers"
-  private var awaitingInput: Boolean = false
 
-  // UI Components
-  private val boardPanel = new BoardPanel
-  private val statusLabel = new Label(statusMessage) {
-    font = new Font("Arial", Font.BOLD, 16)
-    horizontalAlignment = Alignment.Center
-    preferredSize = new Dimension(600, 30)
+  def run(controller: IController): Unit = {
+    CheckersGuiApp.inputHandler = inputHandler
+    CheckersGuiApp.controller = controller
+    CheckersGuiApp.guiInstance = this
+    CheckersGuiApp.main(Array())
   }
 
-  private val scoreLabel = new Label("Red: 12 | Black: 12") {
-    font = new Font("Arial", Font.PLAIN, 14)
-    horizontalAlignment = Alignment.Center
-  }
-
-  private val rulesButton = new Button("Rules (R)") { enabled = true }
-  private val undoButton = new Button("Undo (U)") { enabled = false }
-  private val redoButton = new Button("Redo (R)") { enabled = false }
-  private val quitButton = new Button("Quit (Q)") { enabled = true }
-  private val startButtonPlaceholder = new Button("Start Game") { enabled = false; visible = false }
-
-  private val frame = new MainFrame {
-    title = "Checkers Game"
-
-    listenTo(this)
-    reactions += {
-      case UIElementResized(_) =>
-        boardPanel.revalidate()
-        boardPanel.repaint()
+  override def update(event: GameEvent): Unit = {
+    Platform.runLater {
+      CheckersGuiApp.updateView(event)
     }
-    contents = new BorderPanel {
-      layout(new BoxPanel(Orientation.Vertical) {
-        contents += statusLabel
-        contents += Swing.VStrut(5)
-      }) = BorderPanel.Position.North
+  }
 
-      layout(boardPanel) = BorderPanel.Position.Center
+  private object CheckersGuiApp extends JFXApp3 {
+    private val NUMBER_OF_PIECES: Int = 12
+    private var redTurn: Boolean = true
 
-      layout(new BoxPanel(Orientation.Vertical) {
-        contents += new BoxPanel(Orientation.Horizontal) {
-          contents += scoreLabel
-          contents += Swing.HGlue
+    var inputHandler: InputHandler = uninitialized
+    var controller: IController = uninitialized
+    var guiInstance: GuiView = uninitialized
+    private var canvas: Canvas = uninitialized
+    private var redPointsLabel: Label = uninitialized
+    private var blackPointsLabel: Label = uninitialized
+    private var turnLabel: Label = uninitialized
+    private var currentBoard: Option[Board] = Some(BoardBuilder(8).withStandardSetup().build())
+    private var selectedSquare: Option[(Int, Int)] = None
+
+    // Animation States
+    private var animatingPiece: Option[(Piece, Double, Double, Double, Double, Double)] = None // Piece, fromCol, fromRow, toCol, toRow, progress
+    private var animationTimeline: Timeline = uninitialized
+
+    override def start(): Unit = {
+      canvas = new Canvas()
+
+      redPointsLabel = new Label("0") {
+        style = "-fx-font-size: 44px; -fx-text-fill: white; -fx-font-weight: bold;"
+      }
+      blackPointsLabel = new Label("0") {
+        style = "-fx-font-size: 44px; -fx-text-fill: white; -fx-font-weight: bold;"
+      }
+      turnLabel = new Label("TURN: RED") {
+        style = "-fx-font-size: 22px; -fx-font-weight: bold; -fx-text-fill: #990000; -fx-effect: dropshadow(three-pass-box, rgba(0,0,0,0.3), 2, 0, 0, 2);"
+      }
+
+      val redScoreCard = new VBox {
+        alignment = Pos.Center
+        minWidth = 140
+        minHeight = 140
+        style = "-fx-background-color: linear-gradient(to bottom right, #cc0000, #800000); -fx-background-radius: 20; -fx-border-color: #660000; -fx-border-width: 4; -fx-border-radius: 20; -fx-effect: dropshadow(gaussian, rgba(0,0,0,0.4), 10, 0, 5, 5);"
+        children = Seq(new Label("RED") {
+          style = "-fx-text-fill: white; -fx-font-weight: bold; -fx-font-size: 14px;"
+        }, redPointsLabel)
+      }
+
+      val blackScoreCard = new VBox {
+        alignment = Pos.Center
+        minWidth = 140
+        minHeight = 140
+        style = "-fx-background-color: linear-gradient(to bottom right, #444444, #111111); -fx-background-radius: 20; -fx-border-color: black; -fx-border-width: 4; -fx-border-radius: 20; -fx-effect: dropshadow(gaussian, rgba(0,0,0,0.4), 10, 0, 5, 5);"
+        children = Seq(new Label("BLACK") {
+          style = "-fx-text-fill: white; -fx-font-weight: bold; -fx-font-size: 14px;"
+        }, blackPointsLabel)
+      }
+
+      val sidePanel = new VBox {
+        alignment = Pos.Center
+        padding = Insets(30)
+        spacing = 35
+        prefWidth = 240
+        style = "-fx-background-color: rgba(220, 220, 220, 0.7); -fx-border-color: #aaaaaa; -fx-border-width: 0 0 0 2;"
+        children = Seq(new Label("SCOREBOARD") {
+          style = "-fx-font-size: 20px; -fx-font-weight: bold;"
+        }, redScoreCard, blackScoreCard, turnLabel)
+      }
+
+      val bottomControls = new HBox {
+        alignment = Pos.Center
+        spacing = 25
+        padding = Insets(15)
+        style = "-fx-background-color: #f4f4f4;"
+        children = Seq(
+          new Button("↩ Undo") {
+            style = "-fx-font-size: 14px;"
+            onAction = _ => inputHandler.submitInput("undo")
+          },
+          new Button("↪ Redo") {
+            style = "-fx-font-size: 14px;"
+            onAction = _ => inputHandler.submitInput("redo")
+          },
+          new Button("ℹ Rules") {
+            style = "-fx-font-size: 14px;"
+            onAction = _ => showRulesDialog()
+          },
+          new Button("❌ Quit") {
+            style = "-fx-font-size: 14px;"
+            onAction = _ => sys.exit(0)
+          }
+        )
+      }
+
+      val centerPane = new StackPane {
+        style = "-fx-background-color: #2e7d32;"
+        children = Seq(canvas)
+      }
+
+      canvas.width <== centerPane.width
+      canvas.height <== centerPane.height
+      canvas.width.onChange(draw())
+      canvas.height.onChange(draw())
+
+      stage = new JFXApp3.PrimaryStage {
+        title = "Checkers"
+        scene = new Scene(950, 750) {
+          root = new BorderPane {
+            center = centerPane
+            right = sidePanel
+            bottom = bottomControls
+          }
         }
-        contents += Swing.VStrut(5)
-        contents += new BoxPanel(Orientation.Horizontal) {
-          contents += Swing.HGlue
-          contents += rulesButton // <--- BUTTON PLACEMENT
-          contents += Swing.HStrut(10)
-          contents += undoButton
-          contents += Swing.HStrut(10)
-          contents += redoButton
-          contents += Swing.HStrut(10)
-          contents += quitButton
-          contents += Swing.HGlue
-        }
-        contents += Swing.VStrut(5)
-        contents += startButtonPlaceholder
-      }) = BorderPanel.Position.South
+      }
+
+      controller.add(guiInstance)
+      setupInput()
+      updateScore()
+      draw()
     }
-    size = new Dimension(600, 700)
-    centerOnScreen()
-  }
 
-  // Event listeners
-  listenTo(boardPanel.mouse.clicks)
-  listenTo(undoButton)
-  listenTo(rulesButton) // <--- LISTEN TO NEW BUTTON
-  listenTo(redoButton)
-  listenTo(quitButton)
-
-  reactions += {
-    case MouseClicked(source, point, _, _, _) =>
-      if (awaitingInput) {
-        handleBoardClick(point.x, point.y)
-      }
-    case ButtonClicked(`rulesButton`) => // <--- HANDLE RULES BUTTON CLICK
-      showRulesDialog()
-    case ButtonClicked(`undoButton`) =>
-      if (awaitingInput) {
-        inputHandler.submitInput("undo")
-        awaitingInput = false
-      }
-    case ButtonClicked(`redoButton`) =>
-      if (awaitingInput) {
-        inputHandler.submitInput("redo")
-        awaitingInput = false
-      }
-    case ButtonClicked(`quitButton`) =>
-      if (awaitingInput) {
-        inputHandler.submitInput("quit")
-        awaitingInput = false
-      } else {
-        frame.dispose()
-        System.exit(0)
-      }
-  }
-
-  def show(): Unit = {
-    frame.visible = true
-  }
-
-  // --- NEW METHOD FOR RULES DIALOG ---
-  private def showRulesDialog(): Unit = {
-    val rules =
-      """
-        |Checkers Rules:
-        |
-        |1. **Movement**: Regular pieces move diagonally forward one square.
-        |2. **Kings**: When a piece reaches the opponent's back row, it becomes a King (♔). Kings can move diagonally forward or backward.
-        |3. **Jumping/Capturing**: You must capture an opponent's piece if possible. Captures are made by jumping diagonally over an opponent's piece to the empty square immediately beyond it.
-        |4. **Multiple Jumps**: If a jump lands you in a position where another jump is immediately available (a "multi-jump"), you must take the subsequent jump in the same turn.
-        |5. **Forced Move**: If any jump is available on the board, you MUST take a jump. You cannot make a non-jump move.
-        |6. **Winning**: The game is won by the player who captures all of the opponent's pieces.
-        """.stripMargin.trim
-
-    // Use Swing's simple dialog utility
-    JOptionPane.showMessageDialog(frame.peer, rules, "Checkers Rules", JOptionPane.INFORMATION_MESSAGE)
-  }
-
-  override def update(event: GameEvent): Unit = event match {
-    case StartGame() =>
-      statusMessage = "Press 'Start Game' to begin"
-      statusLabel.text = statusMessage
-      statusLabel.background = new Color(240, 240, 240)
-      statusLabel.opaque = true
-      show()
-
-    case BoardUpdated(board, redTurn) =>
-      Swing.onEDT {
+    def updateView(event: GameEvent): Unit = event match {
+      case BoardUpdated(board, isRed) =>
+        redTurn = isRed
+        detectMove(board)
         currentBoard = Some(board)
-        isRedTurn = redTurn
-        selectedPiece = None
-        validMoves = Nil
-        updateScoreLabel(board)
-        boardPanel.repaint()
-        undoButton.enabled = true
-        redoButton.enabled = true
-      }
+        selectedSquare = None
 
-    case TurnAnnounced(redTurn) =>
-      Swing.onEDT {
-        isRedTurn = redTurn
-        val player = if (redTurn) "Red (●)" else "Black (○)"
-        statusMessage = s"$player's turn - Click a piece to move"
-        statusLabel.text = statusMessage
-        statusLabel.background = if (redTurn) new Color(255, 200, 200) else new Color(200, 200, 200)
-        statusLabel.opaque = true
-      }
-
-    case RequestInput(redTurn) =>
-      Swing.onEDT {
-        awaitingInput = true
-        val player = if (redTurn) "Red" else "Black"
-        if (currentBoard.isEmpty) {
-          statusMessage = "Welcome! Press Start Game to begin"
-
-        } else {
-          statusMessage = s"$player's turn - Make your move"
+        val stackPane = canvas.parent.value.asInstanceOf[javafx.scene.layout.StackPane]
+        val overlays = stackPane.getChildren.filter(_.isInstanceOf[javafx.scene.layout.VBox])
+        if (overlays.nonEmpty) {
+          stackPane.getChildren.removeAll(overlays)
         }
-        statusLabel.text = statusMessage
+        if (redPointsLabel != null) updateScore()
+        draw()
+
+      case GameEnded(finalBoard, winnerIsRed) =>
+        detectMove(finalBoard)
+        currentBoard = Some(finalBoard)
+        updateScore()
+        draw()
+
+        val pause = new scalafx.animation.PauseTransition(scalafx.util.Duration(500))
+        pause.onFinished = _ => {
+          showWinnerOverlay(if (winnerIsRed) "RED" else "BLACK")
+        }
+        pause.play()
+      case _ =>
+    }
+
+    private def detectMove(newBoard: Board): Unit = {
+      for (oldB <- currentBoard) {
+        var from: Option[(Int, Int)] = None
+        var to: Option[(Int, Int)] = None
+        var movedPiece: Option[Piece] = None
+
+        for (r <- 0 until 8; c <- 0 until 8) {
+          if (oldB(r)(c) != Empty && newBoard(r)(c) == Empty) from = Some((r, c))
+          if (oldB(r)(c) == Empty && newBoard(r)(c) != Empty) {
+            to = Some((r, c))
+            movedPiece = Some(newBoard(r)(c))
+          }
+        }
+
+        for ((fR, fC) <- from; (tR, tC) <- to; p <- movedPiece) {
+          animatingPiece = Some((p, fC.toDouble, fR.toDouble, tC.toDouble, tR.toDouble, 0.0))
+          if (animationTimeline != null) animationTimeline.stop()
+
+          animationTimeline = new Timeline {
+            keyFrames = Seq(
+              KeyFrame(Duration(16), onFinished = _ => {
+                animatingPiece = animatingPiece.map(ap => ap.copy(_6 = Math.min(1.0, ap._6 + 0.05)))
+                draw()
+                if (animatingPiece.exists(_._6 >= 1.0)) {
+                  animatingPiece = None
+                  draw()
+                  stop()
+                }
+              })
+            )
+            cycleCount = Timeline.Indefinite
+          }
+          animationTimeline.play()
+        }
+      }
+    }
+
+    private def updateScore(): Unit = {
+      val (red, black) = GameLogic.countPieces(currentBoard.get)
+      redPointsLabel.text = s"${NUMBER_OF_PIECES - black}"
+      blackPointsLabel.text = s"${NUMBER_OF_PIECES - red}"
+      turnLabel.text = if (redTurn) "TURN: RED" else "TURN: BLACK"
+      turnLabel.style = s"-fx-font-size: 22px; -fx-font-weight: bold; -fx-text-fill: ${if (redTurn) "#990000" else "#323232"};"
+    }
+
+    // Helper to keep your 3D piece logic exactly as it was
+    private def drawPiece(gc: scalafx.scene.canvas.GraphicsContext, p: Piece, x: Double, y: Double, tileSize: Double, redTurn: Boolean, board: Option[Board], row: Int, col: Int): Unit = {
+      val isRed = p match {
+        case Regular(r) => r
+        case King(r) => r
+        case _ => false
       }
 
-    case GameEnded(winnerIsRed) =>
-      Swing.onEDT {
-        val winner = if (winnerIsRed) "Red" else "Black"
-        statusMessage = s"🎉 Game Over! $winner wins! 🎉"
-        statusLabel.text = statusMessage
-        statusLabel.background = if (winnerIsRed) new Color(255, 180, 180) else new Color(180, 180, 180)
-        statusLabel.opaque = true
-        selectedPiece = None
-        validMoves = Nil
-        boardPanel.repaint()
-        awaitingInput = false
-        undoButton.enabled = false
-        redoButton.enabled = false
+      // Glow logic (only if row/col provided - skipped for sliding ghost)
+      if (row != -1 && isRed == redTurn) {
+        board.foreach { b =>
+          val moves = GameLogic.getValidMoves(b, row, col)
+          if (if (GameLogic.hasJumpsAvailable(b, redTurn)) moves.exists(_._3) else moves.nonEmpty) {
+            gc.setStroke(Color.web("#FFFFFF", 0.9))
+            gc.setLineWidth(4)
+            gc.strokeRoundRect(x + 4, y + 4, tileSize - 8, tileSize - 8, 12, 12)
+          }
+        }
       }
 
-    case MoveFailed(reason) =>
-      Swing.onEDT {
-        statusMessage = s"❌ Invalid: $reason"
-        statusLabel.text = statusMessage
-        statusLabel.background = new Color(255, 220, 220)
-        statusLabel.opaque = true
-        awaitingInput = true
+      // Piece Shadow (3D Depth)
+      gc.fill = Color.web("#000000", 0.4)
+      gc.fillOval(x + tileSize * 0.18, y + tileSize * 0.18, tileSize * 0.68, tileSize * 0.68)
+
+      // Glossy Body
+      val baseColor = if (isRed) Color.web("#cc0000") else Color.web("#222222")
+      val grad = new LinearGradient(0, 0, 1, 1, true, CycleMethod.NoCycle,
+        List(Stop(0.0, baseColor.brighter), Stop(0.6, baseColor), Stop(1.0, baseColor.darker)))
+
+      gc.fill = grad
+      gc.fillOval(x + tileSize * 0.15, y + tileSize * 0.15, tileSize * 0.7, tileSize * 0.7)
+
+      // Piece Ridge (Ring)
+      gc.setStroke(baseColor.darker)
+      gc.setLineWidth(1.5)
+      gc.strokeOval(x + tileSize * 0.25, y + tileSize * 0.25, tileSize * 0.5, tileSize * 0.5)
+
+      if (p.isInstanceOf[King]) {
+        gc.fill = Color.Gold
+        gc.font = scalafx.scene.text.Font.font("Serif", scalafx.scene.text.FontWeight.Bold, tileSize * 0.5)
+        gc.textAlign = scalafx.scene.text.TextAlignment.Center
+        gc.textBaseline = scalafx.geometry.VPos.Center
+        gc.fillText("♔", x + tileSize / 2, y + tileSize / 2 - (tileSize * 0.05))
+        gc.textAlign = scalafx.scene.text.TextAlignment.Left
+        gc.textBaseline = scalafx.geometry.VPos.Baseline
+      }
+    }
+
+    private def draw(): Unit = {
+      if (canvas == null || canvas.width.value <= 0) return
+      val gc = canvas.graphicsContext2D
+      val size = Math.min(canvas.width.value, canvas.height.value) * 0.9
+      val tileSize = size / 8
+      val offsetX = (canvas.width.value - size) / 2
+      val offsetY = (canvas.height.value - size) / 2
+
+      gc.clearRect(0, 0, canvas.width.value, canvas.height.value)
+
+      // 1. 3D Board Frame
+      gc.fill = Color.web("#5d3a1a")
+      gc.fillRoundRect(offsetX - 10, offsetY - 10, size + 20, size + 20, 10, 10)
+
+      // 2. Main Board Squares
+      for (row <- 0 until 8; col <- 0 until 8) {
+        val x = offsetX + col * tileSize
+        val y = offsetY + row * tileSize
+        gc.fill = if ((row + col) % 2 == 0) Color.web("#ebcd99") else Color.web("#966333")
+        gc.fillRect(x, y, tileSize, tileSize)
+        gc.stroke = Color.web("#000000", 0.1)
+        gc.strokeRect(x, y, tileSize, tileSize)
       }
 
-    case InvalidInput(message) =>
-      Swing.onEDT {
-        statusMessage = s"❌ Error: $message"
-        statusLabel.text = statusMessage
-        statusLabel.background = new Color(255, 220, 220)
-        statusLabel.opaque = true
-        awaitingInput = true
+      // 3. Move Predictions
+      selectedSquare.foreach { case (sR, sC) =>
+        currentBoard.foreach { b =>
+          val moves = GameLogic.getValidMoves(b, sR, sC)
+          val hasJ = GameLogic.hasJumpsAvailable(b, redTurn)
+          (if (hasJ) moves.filter(_._3) else moves).foreach { case (mR, mC, isJump) =>
+            gc.fill = if (isJump) Color.web("#8B0000", 0.5) else Color.web("#00FF00", 0.3)
+            gc.fillRect(offsetX + mC * tileSize, offsetY + mR * tileSize, tileSize, tileSize)
+          }
+        }
       }
 
-    case KillEffect(kills) =>
-      Swing.onEDT {
-        statusMessage = s"💥 ${kills} piece${if (kills > 1) "s" else ""} captured!"
-        statusLabel.text = statusMessage
-        statusLabel.background = new Color(255, 255, 180)
-        statusLabel.opaque = true
+      // 4. Static Pieces
+      for (row <- 0 until 8; col <- 0 until 8) {
+        currentBoard.foreach { b =>
+          val p = b(row)(col)
+          val isArriving = animatingPiece.exists(ap => ap._4.toInt == col && ap._5.toInt == row)
+          if (p != Empty && !isArriving) {
+            drawPiece(gc, p, offsetX + col * tileSize, offsetY + row * tileSize, tileSize, redTurn, currentBoard, row, col)
+          }
+        }
       }
 
-    case MoveUndone() =>
-      Swing.onEDT {
-        statusMessage = "⬅️ Move undone"
-        statusLabel.text = statusMessage
-        statusLabel.background = new Color(220, 220, 255)
-        statusLabel.opaque = true
+      // 5. Sliding Piece
+      animatingPiece.foreach { case (p, fC, fR, tC, tR, progress) =>
+        val curC = fC + (tC - fC) * progress
+        val curR = fR + (tR - fR) * progress
+        drawPiece(gc, p, offsetX + curC * tileSize, offsetY + curR * tileSize, tileSize, redTurn, None, -1, -1)
       }
+    }
 
-    case MoveRedone() =>
-      Swing.onEDT {
-        statusMessage = "➡️ Move redone"
-        statusLabel.text = statusMessage
-        statusLabel.background = new Color(220, 255, 220)
-        statusLabel.opaque = true
-      }
+    private def setupInput(): Unit = {
+      canvas.onMouseClicked = e => {
+        val size = Math.min(canvas.width.value, canvas.height.value) * 0.9
+        val offsetX = (canvas.width.value - size) / 2
+        val offsetY = (canvas.height.value - size) / 2
+        val col = ((e.x - offsetX) / (size / 8)).toInt
+        val row = ((e.y - offsetY) / (size / 8)).toInt
 
-    case QuitGame() =>
-      Swing.onEDT {
-        frame.dispose()
-        System.exit(0)
-      }
-  }
-
-  private def updateScoreLabel(board: Board): Unit = {
-    val (red, black) = GameLogic.countPieces(board)
-    scoreLabel.text = s"Red: $red | Black: $black"
-  }
-
-  // --- FIX 1: Remove all rotation logic from click handling ---
-  // --- In GuiView.scala (Modified handleBoardClick) ---
-
-  // This method must be synchronous and fast, running on the EDT
-  private def handleBoardClick(x: Int, y: Int): Unit = {
-    currentBoard.foreach { board =>
-
-      // 1. Get the actual, current size of the *square* board area.
-      // This MUST match the calculation used in paintComponent.
-      val boardSize = math.min(boardPanel.size.width, boardPanel.size.height)
-
-      // 2. Calculate the correct, current cell size.
-      val cellSize = boardSize / 8
-
-      // 3. Since the board is centered, we might need to adjust coordinates
-      //    if the click is in the empty margins.
-      //    If the click is outside the boardSize x boardSize area, ignore it.
-      if (x >= 0 && x < boardSize && y >= 0 && y < boardSize) {
-
-        // 4. Calculate coordinates using the new cellSize.
-        val internalCol = x / cellSize
-        val internalRow = y / cellSize
-
-        if (internalRow >= 0 && internalRow < 8 && internalCol >= 0 && internalCol < 8) {
-
-          selectedPiece match {
-            case Some((selRow, selCol)) =>
-              if (validMoves.exists { case (r, c, _) => r == internalRow && c == internalCol }) {
-                executeMove(selRow, selCol, internalRow, internalCol)
-              } else {
-                trySelectPiece(board, internalRow, internalCol)
-              }
-
+        if (col >= 0 && col < 8 && row >= 0 && row < 8) {
+          selectedSquare match {
             case None =>
-              trySelectPiece(board, internalRow, internalCol)
-          }
-        }
-      }
-    }
-  }
-
-  private def trySelectPiece(board: Board, row: Int, col: Int): Unit = {
-    Swing.onEDT {
-      board(row)(col) match {
-        case Regular(red) if red == isRedTurn =>
-          selectPiece(board, row, col)
-        case King(red) if red == isRedTurn =>
-          selectPiece(board, row, col)
-        case _ =>
-          selectedPiece = None
-          validMoves = Nil
-          statusMessage = "Select one of your own pieces"
-          statusLabel.text = statusMessage
-          boardPanel.repaint()
-      }
-    }
-  }
-
-  private def selectPiece(board: Board, row: Int, col: Int): Unit = {
-    Swing.onEDT {
-      val moves = GameLogic.getValidMoves(board, row, col)
-      val hasJumps = GameLogic.hasJumpsAvailable(board, isRedTurn)
-
-      validMoves = if (hasJumps) moves.filter(_._3) else moves
-
-      if (validMoves.isEmpty) {
-        statusMessage = if (hasJumps) "❌ Must jump! Select a piece that can jump." else "❌ No valid moves for this piece"
-        statusLabel.text = statusMessage
-        statusLabel.background = new Color(255, 220, 220)
-        statusLabel.opaque = true
-        selectedPiece = None
-      } else {
-        selectedPiece = Some((row, col))
-        val colChar = ('a' + col).toChar
-        val rowNum = row + 1
-        statusMessage = s"✓ Selected ($colChar$rowNum). Click destination (Green=move, Red=jump)"
-        statusLabel.text = statusMessage
-        statusLabel.background = new Color(220, 255, 220)
-        statusLabel.opaque = true
-      }
-
-      boardPanel.repaint()
-    }
-  }
-
-  private def executeMove(fromRow: Int, fromCol: Int, toRow: Int, toCol: Int): Unit = {
-    val fromColChar = ('a' + fromCol).toChar
-    val fromRowNum = fromRow + 1
-    val toColChar = ('a' + toCol).toChar
-    val toRowNum = toRow + 1
-
-    val input = s"$fromColChar$fromRowNum $toColChar$toRowNum"
-
-    inputHandler.submitInput(input)
-
-    Swing.onEDT {
-      awaitingInput = false
-      selectedPiece = None
-      validMoves = Nil
-      boardPanel.repaint()
-    }
-  }
-
-  // Custom panel for drawing the checkers board
-  private class BoardPanel extends Panel {
-    preferredSize = new Dimension(600, 600)
-
-    override def preferredSize: Dimension = {
-      // If the frame has not been packed/shown yet, return a default minimum.
-      if (frame.contents.head.size.width == 0) new Dimension(600, 600)
-      else {
-        // Get the current available width and height for the central panel
-        val containerWidth = frame.contents.head.size.width
-        val containerHeight = frame.contents.head.size.height - statusLabel.preferredSize.height -
-          scoreLabel.preferredSize.height - 40 // ~40 accounts for padding/gaps/buttons
-
-        // The board size is the minimum of the available width and height
-        val size = math.min(containerWidth, containerHeight)
-        new Dimension(size, size)
-      }
-    }
-
-    override def paintComponent(g: Graphics2D): Unit = {
-      val boardSize = math.min(size.width, size.height)
-      val cellSize = boardSize / 8
-
-      super.paintComponent(g)
-      g.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON)
-
-      currentBoard.foreach { board =>
-        val hasJumpsAvailable = GameLogic.hasJumpsAvailable(board, isRedTurn)
-
-        // Draw board squares, coordinates, and highlights
-        for (displayRow <- 0 until 8; displayCol <- 0 until 8) {
-          val x = displayCol * cellSize
-          val y = displayRow * cellSize
-
-          // --- FIX 2 & 3: Internal coordinates are now ALWAYS the same as Display Coordinates ---
-          val internalRow = displayRow
-          val internalCol = displayCol
-
-          // Checkerboard pattern
-          g.setColor(if ((displayRow + displayCol) % 2 == 0) new Color(240, 217, 181) else new Color(181, 136, 99))
-          g.fillRect(x, y, cellSize, cellSize)
-
-          // Draw coordinate labels (Fixed: Always display 1-8 from bottom to top, a-h from left to right)
-          // Row numbers are displayed according to the internal coordinate (0=top row/Black home, 7=bottom row/Red home)
-          val rowLabel = 8 - displayRow
-          if (displayCol == 0) {
-            g.setColor(Color.BLACK)
-            g.setFont(new Font("Arial", Font.PLAIN, 10))
-            g.drawString(s"$rowLabel", x + 2, y + 12)
-          }
-          if (displayRow == 7) {
-            g.setColor(Color.BLACK)
-            g.setFont(new Font("Arial", Font.PLAIN, 10))
-            val colLabel = ('a' + displayCol).toChar
-            g.drawString(colLabel.toString, x + cellSize - 12, y + cellSize - 2)
-          }
-
-          // --- Playable Piece Highlighting (using internalRow/Col which are now fixed) ---
-          board(internalRow)(internalCol) match {
-            case Regular(isRed) if isRed == isRedTurn =>
-              handleHighlight(g, board, internalRow, internalCol, hasJumpsAvailable, x, y, cellSize)
-            case King(isRed) if isRed == isRedTurn =>
-              handleHighlight(g, board, internalRow, internalCol, hasJumpsAvailable, x, y, cellSize)
-            case _ => // Empty or opponent's piece: no highlighting
-          }
-
-          // Highlight selected piece
-          selectedPiece.foreach { case (selRow, selCol) =>
-            if (selRow == internalRow && selCol == internalCol) {
-              g.setColor(new Color(255, 255, 0, 120))
-              g.fillRect(x, y, cellSize, cellSize)
-              g.setStroke(new BasicStroke(3))
-              g.setColor(new Color(255, 215, 0))
-              g.drawRect(x + 2, y + 2, cellSize - 4, cellSize - 4)
-            }
-          }
-
-          // Highlight valid moves
-          validMoves.foreach { case (moveRow, moveCol, isJump) =>
-            if (moveRow == internalRow && moveCol == internalCol) {
-              g.setColor(if (isJump) new Color(255, 0, 0, 100) else new Color(0, 255, 0, 100))
-              g.fillRect(x, y, cellSize, cellSize)
-
-              // Draw indicator circle
-              g.setColor(if (isJump) new Color(200, 0, 0) else new Color(0, 180, 0))
-              g.setStroke(new BasicStroke(3))
-              val circleSize = cellSize / 3
-              g.drawOval(x + cellSize/2 - circleSize/2, y + cellSize/2 - circleSize/2, circleSize, circleSize)
-            }
-          }
-
-          // Draw pieces (last, to be on top of all highlights)
-          board(internalRow)(internalCol) match {
-            case Regular(isRed) =>
-              drawPiece(g, x, y, cellSize, isRed, isKing = false)
-            case King(isRed) =>
-              drawPiece(g, x, y, cellSize, isRed, isKing = true)
-            case Empty =>
+              currentBoard.foreach { b =>
+                val piece = b(row)(col)
+                val isValidSelection = piece match {
+                  case Regular(isRed) => isRed == redTurn
+                  case King(isRed) => isRed == redTurn
+                  case _ => false
+                }
+                if (isValidSelection) {
+                  selectedSquare = Some((row, col))
+                  draw()
+                }
+              }
+            case Some((sR, sC)) =>
+              if (sR == row && sC == col) {
+                selectedSquare = None
+                draw()
+              } else {
+                val moveCommand = s"${('a' + sC).toChar}${sR + 1} ${('a' + col).toChar}${row + 1}"
+                try {
+                  inputHandler.submitInput(moveCommand)
+                } catch {
+                  case ex: Exception => ex.printStackTrace()
+                }
+                selectedSquare = None
+                draw()
+              }
           }
         }
       }
     }
 
-    private def handleHighlight(g: Graphics2D, board: Board, row: Int, col: Int, hasJumpsAvailable: Boolean, x: Int, y: Int, cellSize: Int): Unit = {
-      val moves = GameLogic.getValidMoves(board, row, col)
-      val isPlayable = if (hasJumpsAvailable) moves.exists(_._3) else moves.nonEmpty
-
-      if (isPlayable) {
-        g.setColor(new Color(0, 150, 255, 100)) // Blue overlay
-        g.fillOval(x + 5, y + 5, cellSize - 10, cellSize - 10)
-
-        g.setStroke(new BasicStroke(2))
-        g.setColor(new Color(0, 100, 255)) // Darker Blue Ring
-        g.drawOval(x + 5, y + 5, cellSize - 10, cellSize - 10)
+    private def showWinnerOverlay(winner: String): Unit = {
+      lazy val overlay: VBox = new VBox {
+        alignment = Pos.Center
+        spacing = 10
+        padding = Insets(50)
+        style =
+          """
+          -fx-background-color: radial-gradient(center 50% 50%, radius 80%, rgba(40, 30, 20, 0.95), rgba(10, 10, 10, 0.98));
+          -fx-background-radius: 20;
+          -fx-border-color: linear-gradient(to bottom, #FFD700, #B8860B, #8B4513);
+          -fx-border-width: 5;
+          -fx-border-radius: 15;
+          -fx-effect: dropshadow(gaussian, gold, 30, 0, 0, 0);
+        """
+        maxWidth = 550
+        maxHeight = 350
+        children = Seq(
+          new Label("✧ VICTORY ✧") {
+            style = "-fx-text-fill: linear-gradient(to bottom, #FFFFFF, #FFD700); -fx-font-size: 32px; -fx-font-weight: bold; -fx-font-family: 'Georgia';"
+          },
+          new Label(s"$winner") {
+            val textColor: String = if (winner.toUpperCase == "RED") "#FF3333" else "#CCCCCC"
+            style = s"-fx-text-fill: $textColor; -fx-font-size: 68px; -fx-font-weight: bold; -fx-font-family: 'Verdana'; -fx-effect: dropshadow(one-pass-box, black, 5, 5, 0, 0);"
+          },
+          new Label("HAS CONQUERED THE BOARD") {
+            style = "-fx-text-fill: #DAA520; -fx-font-size: 16px; -fx-font-letter-spacing: 2px; -fx-font-weight: bold;"
+          },
+          new HBox {
+            alignment = Pos.Center
+            padding = Insets(30, 0, 0, 0)
+            children = Seq(
+              new Button("PLAY AGAIN") {
+                style =
+                  """
+                      -fx-background-color: linear-gradient(#2e7d32, #1b5e20);
+                      -fx-text-fill: white;
+                      -fx-font-size: 18px;
+                      -fx-font-weight: bold;
+                      -fx-padding: 12 30;
+                      -fx-background-radius: 5;
+                      -fx-border-color: #81c784;
+                      -fx-border-width: 2;
+                      -fx-cursor: hand;
+                    """
+                onAction = _ => {
+                  de.htwg.controller.command.CommandHistory.clear()
+                  inputHandler.submitInput("revanche")
+                }
+              },
+              new Button("QUIT") {
+                style =
+                  """
+                      -fx-background-color: linear-gradient(#4b3621, #2c1e12);
+                      -fx-text-fill: #DAA520;
+                      -fx-font-size: 18px;
+                      -fx-font-weight: bold;
+                      -fx-padding: 12 30;
+                      -fx-background-radius: 5;
+                      -fx-border-color: #DAA520;
+                      -fx-border-width: 2;
+                      -fx-cursor: hand;
+                    """
+                onAction = _ => sys.exit(0)
+              }
+            )
+          }
+        )
       }
+      val stackPane = canvas.parent.value.asInstanceOf[javafx.scene.layout.StackPane]
+      stackPane.getChildren.add(overlay)
+      overlay.opacity = 0.0
+      overlay.scaleX = 0.5
+      overlay.scaleY = 0.5
+      val fadeIn = new scalafx.animation.FadeTransition(scalafx.util.Duration(600), overlay) {
+        toValue = 1.0
+      }
+      val scaleUp = new scalafx.animation.ScaleTransition(scalafx.util.Duration(600), overlay) {
+        toX = 1.0
+        toY = 1.0
+      }
+      new scalafx.animation.ParallelTransition(Seq(fadeIn, scaleUp)).play()
     }
 
-    private def drawPiece(g: Graphics2D, x: Int, y: Int, size: Int, isRed: Boolean, isKing: Boolean): Unit = {
-      val pieceSize = (size * 0.65).toInt
-      val offset = (size - pieceSize) / 2
-      val centerX = x + offset
-      val centerY = y + offset
-
-      // Draw shadow
-      g.setColor(new Color(0, 0, 0, 60))
-      g.fillOval(centerX + 3, centerY + 3, pieceSize, pieceSize)
-
-      // Draw piece main color
-      g.setColor(if (isRed) new Color(200, 30, 30) else new Color(50, 50, 50))
-      g.fillOval(centerX, centerY, pieceSize, pieceSize)
-
-      // Draw highlight
-      g.setColor(new Color(255, 255, 255, 80))
-      g.fillOval(centerX + pieceSize/4, centerY + pieceSize/6, pieceSize/3, pieceSize/4)
-
-      // Draw border
-      g.setStroke(new BasicStroke(2))
-      g.setColor(if (isRed) new Color(150, 0, 0) else Color.BLACK)
-      g.drawOval(centerX, centerY, pieceSize, pieceSize)
-
-      // Draw king crown
-      if (isKing) {
-        g.setColor(new Color(255, 215, 0))
-        g.setFont(new Font("Serif", Font.BOLD, (size * 0.5).toInt))
-        val metrics = g.getFontMetrics
-        val crownText = "♔"
-        val crownX = centerX + (pieceSize - metrics.stringWidth(crownText)) / 2
-        val crownY = centerY + ((pieceSize - metrics.getHeight) / 2) + metrics.getAscent
-        g.drawString(crownText, crownX, crownY)
+    private def showRulesDialog(): Unit = {
+      val alert = new Alert(AlertType.Information) {
+        initOwner(stage)
+        title = "*** Checkers Rules ***"
+        headerText = "Game Rules:"
       }
+      val content = new Label {
+        text = "1. Movement: Diagonally forward.\n2. Kings: Move diagonally forward/backward.\n3. Jumps: Forced if available.\n4. Winning: Capture all opponent pieces."
+        style = "-fx-font-size: 14px; -fx-padding: 10;"
+      }
+      alert.getDialogPane.setContent(content)
+      alert.getDialogPane.setPrefWidth(500)
+      alert.showAndWait()
     }
   }
 }
